+++ src/app.ts
@@ -9,7 +9,8 @@ import lineReader from 'line-reader';
 export interface Args {
   exclude?: string[];
   files: string;
-  'luhn-check'?: boolean;
+  ignoreNumbers?: string[];
+  luhnCheck?: boolean;
   silent?: boolean;
   verbose?: boolean;
 }
@@ -17,12 +18,17 @@ export interface Args {
 const defaultArgs = {
   exclude: [],
   files: '**/*.{ts,js}',
-  'luhn-check': true,
+  ignoreNumbers: [],
+  luhnCheck: true,
   silent: false,
   verbose: false
 };
 
-const isCardNumber = (suspect: string): boolean => luhn(suspect.replace(/[\s-]+/g, ''));
+const isCardNumber = (suspect: string, ignoreNumbers?: string[]): boolean => {
+  const trimmedSuspect = suspect.replace(/[\s-]+/g, '');
+
+  return !ignoreNumbers?.includes(trimmedSuspect) && luhn(trimmedSuspect);
+}
 
 const showMatches = async (
   fileName: string,
@@ -43,15 +49,15 @@ const showMatches = async (
         if (line.match(new RegExp(`.*${match}.*`))) {
           const columnNumber = line.search(match) + 1;
 
-          if (args['luhn-check'] && isCardNumber(match)) {
+          if (args.luhnCheck && isCardNumber(match, args.ignoreNumbers)) {
             error = true;
             outputLines.push(
               `${chalk.gray(`${lineNumber}:${columnNumber}`.padEnd(8))} ${chalk.red(
                 'error'.padEnd(6)
               )} ${line.replace(match, chalk.red(match))}`
             );
-          } else if (!args['luhn-check']) {
-            if (isCardNumber(match)) {
+          } else if (!args.luhnCheck) {
+            if (isCardNumber(match, args.ignoreNumbers)) {
               error = true;
               outputLines.push(
                 `${chalk.gray(`${lineNumber}:${columnNumber}`.padEnd(8))} ${chalk.red(
@@ -73,7 +79,7 @@ const showMatches = async (
       }
 
       if (last) {
-        if ((!args.silent && outputLines.length > 0 && error) || (warning && !args['luhn-check'])) {
+        if ((!args.silent && outputLines.length > 0 && error) || (warning && !args.luhnCheck)) {
           console.log(fileName);
           console.log();
           outputLines.forEach((line: string): void => {
@@ -90,14 +96,14 @@ const showMatches = async (
 
 const scanFile = async (fileName: string, args: Args): Promise<boolean> => {
   const contents = fs.readFileSync(fileName, 'utf-8');
-  const matches = args['luhn-check']
+  const matches = args.luhnCheck
     ? contents.match(/\b[2-6]{1}\d{3}[\s-]*\d{4}[\s-]*\d{4}[\s-]*\d{4}[\s-]*\b/gm)
     : contents.match(/\b[1-9]{1}\d{3}[\s-]*\d{4}[\s-]*\d{4}[\s-]*\d{4}[\s-]*\b/gm);
 
   if (matches && matches.length > 0) {
     !args.silent && (await showMatches(fileName, matches, args));
 
-    if (!args['luhn-check'] || matches.filter(luhn).length > 0) {
+    if (!args.luhnCheck || matches.filter(luhn).length > 0) {
       return true;
     }
   }
@@ -160,6 +166,11 @@ const run = async (): Promise<void> => {
     .option('exclude', {
       default: [],
       describe: 'exclude pattern',
+      type: 'array'
+    })
+    .option('ignore-numbers', {
+      default: [],
+      describe: 'ignore these numbers',
       type: 'array'
     })
     .option('luhn-check', {